requirements.txt
aiogram>=3.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
gunicorn>=21.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pydantic>=2.4.0
jinja2>=3.1.2
python-multipart>=0.0.6
asgiref>=3.7.2
aiohttp>=3.8.5